Sync works correctly now + added git_cmd wrapper
The addition of a new repository, working with remoets and syncing
between upstream and downstream repositories is working now and should
be fully functional.

Since there were many git_cmd calls in the bare cache repository, the
git_bare_cmd command is a wrapper that automatically forms the git
command in a bare repository, with the feature that the command can
also be given as string instead of as list, to increase readability.

Signed-off-by: Alin Dobre <alinmd@gmail.com>